dps_django/requirements.txt
ngo==1.8.6
MySQL-python==1.2.5
django-admin-view-permission==0.6
sendgrid==3.6.3
openpyxl==2.5.0a2

